+++ bittorrent/src/test/java/org/johnnei/javatorrent/torrent/files/BlockTest.java
@@ -2,9 +2,7 @@ package org.johnnei.javatorrent.torrent.files;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * Tests {@link Block}
@@ -20,6 +18,18 @@ public class BlockTest {
 			() -> assertEquals(24, block.getSize(), "Incorrect size"),
 			() -> assertEquals(BlockStatus.Needed, block.getStatus(), "Incorrect status"),
 			() -> assertTrue(block.toString().startsWith("Block["), "Incorrect toString start")
+		);
+	}
+
+	@Test
+	public void testCopyWithStatus() {
+		Block base = new Block(3, 15);
+		Block request = Block.copyWithStatus(base, BlockStatus.Requested);
+		assertAll(
+			() -> assertNotEquals(base, request),
+			() -> assertEquals(request.getIndex(), 3),
+			() -> assertEquals(request.getSize(), 15),
+			() -> assertEquals(request.getStatus(), BlockStatus.Requested)
 		);
 	}
 }