+++ bittorrent/src/test/java/org/johnnei/javatorrent/torrent/peer/PeerTest.java
@@ -37,10 +37,7 @@ import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.eq;
 import static org.mockito.Matchers.isA;
 import static org.mockito.Matchers.isNotNull;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 public class PeerTest {
 
@@ -224,10 +221,32 @@ public class PeerTest {
 		when(pieceMock.getFileSet()).thenReturn(fileSetMock);
 
 		Peer peer = DummyEntity.createPeer(socketMock, torrentMock);
+		peer.setChoked(PeerDirection.Download, false);
 		peer.addBlockRequest(pieceMock, 15, 15, PeerDirection.Download);
 
 		assertEquals(1, peer.getWorkQueueSize(PeerDirection.Download), "Working queue should have increased");
 		verify(socketMock).enqueueMessage(any());
+	}
+
+	@Test
+	public void testRejectAddBlockRequestDownloadWhenChoked() {
+		Torrent torrentMock = mock(Torrent.class);
+		BitTorrentSocket socketMock = mock(BitTorrentSocket.class);
+		TorrentFileSet fileSetMock = mock(TorrentFileSet.class);
+		Piece pieceMock = mock(Piece.class);
+
+		when(fileSetMock.getBlockSize()).thenReturn(15);
+		when(fileSetMock.getBitfieldBytes()).thenReturn(new byte[1]);
+		when(fileSetMock.getRequestFactory()).thenReturn(mock(TorrentFileSetRequestFactory.class));
+		when(pieceMock.getIndex()).thenReturn(0);
+		when(pieceMock.getFileSet()).thenReturn(fileSetMock);
+
+		Peer peer = DummyEntity.createPeer(socketMock, torrentMock);
+		peer.setChoked(PeerDirection.Download, true);
+		assertFalse(peer.addBlockRequest(pieceMock, 15, 15, PeerDirection.Download));
+
+		assertEquals(0, peer.getWorkQueueSize(PeerDirection.Download), "Working queue should have increased");
+		verify(socketMock, never()).enqueueMessage(any());
 	}
 
 	@Test
@@ -242,6 +261,7 @@ public class PeerTest {
 		when(fileSetMock.getBitfieldBytes()).thenReturn(new byte[1]);
 
 		Peer peer = DummyEntity.createPeer(socketMock, torrentMock);
+		peer.setChoked(PeerDirection.Upload, false);
 		peer.addBlockRequest(pieceMock, 15, 15, PeerDirection.Upload);
 
 		assertEquals(1, peer.getWorkQueueSize(PeerDirection.Upload), "Working queue should have increased");
@@ -266,6 +286,7 @@ public class PeerTest {
 		when(fileSetMock.getRequestFactory()).thenReturn(requestFactoryMock);
 
 		Peer peer = DummyEntity.createPeer(socketMock, torrentMock);
+		peer.setChoked(PeerDirection.Download, false);
 		peer.addBlockRequest(pieceMock, 15, 15, PeerDirection.Download);
 		peer.addBlockRequest(pieceMock, 30, 15, PeerDirection.Download);
 		assertEquals(2, peer.getWorkQueueSize(PeerDirection.Download), "Working queue should have two items");
@@ -293,6 +314,7 @@ public class PeerTest {
 		when(pieceMock.getFileSet()).thenReturn(fileSetMock);
 
 		Peer peer = DummyEntity.createPeer(socketMock, torrentMock);
+		peer.setChoked(PeerDirection.Upload, false);
 		peer.addBlockRequest(pieceMock, 15, 15, PeerDirection.Upload);
 		peer.addBlockRequest(pieceMock, 30, 15, PeerDirection.Upload);
 		assertEquals(2, peer.getWorkQueueSize(PeerDirection.Upload), "Working queue should have two items");
@@ -320,6 +342,7 @@ public class PeerTest {
 		when(fileSetMock.getRequestFactory()).thenReturn(requestFactoryMock);
 
 		Peer peer = DummyEntity.createPeer(socketMock, torrentMock);
+		peer.setChoked(PeerDirection.Download, false);
 		peer.addBlockRequest(pieceMock, 15, 15, PeerDirection.Download);
 		peer.addBlockRequest(pieceMock, 30, 15, PeerDirection.Download);
 		assertEquals(2, peer.getWorkQueueSize(PeerDirection.Download), "Working queue should have two items");
@@ -367,6 +390,7 @@ public class PeerTest {
 				.setExtensionBytes(DummyEntity.createRandomBytes(8))
 				.build();
 
+		cut.setChoked(PeerDirection.Upload, false);
 		cut.addBlockRequest(pieceMock, 0, 15, PeerDirection.Upload);
 		cut.queueNextPieceForSending();
 		cut.queueNextPieceForSending();
@@ -395,6 +419,7 @@ public class PeerTest {
 				.setExtensionBytes(DummyEntity.createRandomBytes(8))
 				.build();
 
+		cut.setChoked(PeerDirection.Upload, false);
 		cut.addBlockRequest(pieceMock, 0, 15, PeerDirection.Upload);
 		cut.queueNextPieceForSending();
 
@@ -426,6 +451,7 @@ public class PeerTest {
 				.setExtensionBytes(DummyEntity.createRandomBytes(8))
 				.build();
 
+		cut.setChoked(PeerDirection.Download, false);
 		cut.addBlockRequest(pieceMock, 0, 15, PeerDirection.Download);
 		cut.addBlockRequest(pieceMock, 15, 15, PeerDirection.Download);
 
@@ -518,6 +544,7 @@ public class PeerTest {
 
 		assertEquals(1, cut.getFreeWorkTime(), "Initial free work time incorrect");
 
+		cut.setChoked(PeerDirection.Download, false);
 		cut.addBlockRequest(pieceMock, 0, 15, PeerDirection.Download);
 
 		assertEquals(0, cut.getFreeWorkTime(), "Work time should have been affected by the download job.");